Some easy loss-of-precision fixes.

+++ src/end.c
@@ -207,7 +207,7 @@ NH_panictrace_libc(void)
 {
 #ifdef PANICTRACE_LIBC
     void *bt[20];
-    size_t count, x;
+    int count, x;
     char **info, buf[BUFSZ];
 
     raw_print("  Generating more information you may report:\n");
